Solution to CoffeeShop exercise
* Uses `require` for internal files: CoffeeShop.js and menu.json
* Includes test() method for demonstrating the other methods

The instructions require methods to _return_ a result, but the
examples imply that the output is shown with console.log(). To
deal with this, the CoffeeShop class includes a show() method
which accepts a command plus any number of parameters for the
command, and which console.log()s the result of the command.